requirements.txt

email-validator>=2.2.0
flask>=3.1.1
flask-sqlalchemy>=3.1.1
gunicorn>=23.0.0
moviepy>=2.2.1
numpy>=2.3.0
openai>=1.86.0
opencv-python>=4.11.0.86
pillow>=11.2.1
psycopg2-binary>=2.9.10
requests>=2.32.4
scipy>=1.15.3
sqlalchemy>=2.0.41
werkzeug>=3.1.3
